Move WPNestedScrollView to the widgets sub-package
