+++ WordPress/src/main/res/layout/media_settings_activity.xml
@@ -72,7 +72,7 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <org.wordpress.android.ui.stats.WPNestedScrollView
+    <org.wordpress.android.widgets.WPNestedScrollView
         android:id="@+id/scroll_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -549,7 +549,7 @@
 
         </RelativeLayout>
 
-    </org.wordpress.android.ui.stats.WPNestedScrollView>
+    </org.wordpress.android.widgets.WPNestedScrollView>
 
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fab_button"
