+++ WordPress/src/main/res/layout/notifications_fragment_notes_list.xml
@@ -206,7 +206,7 @@
     </LinearLayout>
 
 
-    <org.wordpress.android.ui.stats.WPNestedScrollView
+    <org.wordpress.android.widgets.WPNestedScrollView
         android:id="@+id/connect_jetpack"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -241,5 +241,5 @@
                 android:text="@string/stats_jetpack_connection_setup"
                 android:theme="@style/JetpackConnectionButton"/>
         </LinearLayout>
-    </org.wordpress.android.ui.stats.WPNestedScrollView>
+    </org.wordpress.android.widgets.WPNestedScrollView>
 </android.support.design.widget.CoordinatorLayout>
