+++ WordPress/src/main/res/layout/plugin_detail_activity.xml
@@ -59,7 +59,7 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <org.wordpress.android.ui.stats.WPNestedScrollView
+    <org.wordpress.android.widgets.WPNestedScrollView
         android:id="@+id/scroll_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -448,5 +448,5 @@
 
             <include layout="@layout/plugin_ratings_cardview" />
         </LinearLayout>
-    </org.wordpress.android.ui.stats.WPNestedScrollView>
+    </org.wordpress.android.widgets.WPNestedScrollView>
 </android.support.design.widget.CoordinatorLayout>
